+++ components/messaging/OpenConversation.tsx
@@ -1,33 +1,48 @@
 'use client';
 
-import {
-  ConversationCardType,
-  MessageType,
-} from '@/utils/messaging/messagingTypes';
+import { MessageType } from '@/utils/messaging/messagingTypes';
 import MessageCard from './MessageCard';
 import MessageForm from './MessageForm';
+import { useContext, useEffect } from 'react';
+import { ConversationContext } from './ConversationProvider';
 
-const OpenConversation: React.FC<ConversationCardType> = ({
-  conversations,
-  user_id,
-  conversation_id,
-}) => {
+const OpenConversation: React.FC = () => {
+  const { allConversations, openConversation, setOpenConversation } =
+    useContext(ConversationContext);
+
+  useEffect(() => {
+    const highestId = Math.max(
+      ...allConversations.map((conversation) => conversation.conversation_id)
+    );
+
+    setOpenConversation &&
+      setOpenConversation(
+        allConversations?.filter(
+          (conversation) => conversation.conversation_id === highestId
+        )[0]
+      );
+  }, []);
+
   return (
     <div className='flex w-2/4 flex-col'>
-      {conversations?.messages?.map((message: MessageType) => (
-        <div key={`${conversations.id}-${message.created_at}`}>
-          <MessageCard
-            sender_id={message.sender_id}
-            created_at={message.created_at}
-            message_text={message.message_text}
-            is_read={message.is_read}
-            currentUser={user_id}
-          />
-        </div>
-      ))}
+      {openConversation?.conversations?.messages?.map(
+        (message: MessageType) => (
+          <div
+            key={`${openConversation.conversations.id}-${message.created_at}`}
+          >
+            <MessageCard
+              sender_id={message.sender_id}
+              created_at={message.created_at}
+              message_text={message.message_text}
+              is_read={message.is_read}
+              currentUser={openConversation.user_id}
+            />
+          </div>
+        )
+      )}
       <MessageForm
-        user_id={user_id}
-        conversation_id={conversation_id}
+        user_id={openConversation?.user_id}
+        conversation_id={openConversation?.conversation_id}
       ></MessageForm>
     </div>
   );
